Get API documentation sample from index
See https://github.com/hbz/lobid-organisations/issues/313

Involved some changes and cleanup to the general index access
