+++ app/views/api.scala.html
@@ -3,6 +3,7 @@
 @(title: String)
 
 @import play.i18n._
+@import play.api.libs.json._
 
 @desc(label: String, link: play.api.mvc.Call) = {
  <dt>@label</dt> <dd><a href="@link">@link.toString.replace("%3A",":").replace("%2C",",").replaceAll("[&?]format=json", "")</a></dd>
@@ -34,74 +35,7 @@
     <dt>@Messages.get("api.jsonld.rdf.header")</dt><dd>@Html(Messages.get("api.jsonld.rdf.text", "https://github.com/digitalbazaar/jsonld-cli"))</dd>
     <p><code>jsonld format --nquads http://lobid.org@routes.Application.get("DE-6")</code></p>
     <dt>@Messages.get("api.jsonld.example.header")</dt><dd>@Html(Messages.get("api.jsonld.example.text", "https://hypothes.is/"))</dd>
-    <dd><pre><code>{
-  "linkedTo" : "http://lobid.org/organisations/DE-605#!",
-  "rs" : "053150000000",
-  "telephone" : "0049 221 4 78-7070",
-  "alternateName" : [ "ZB MED - Leibniz-Informationszentrum Lebenswissenschaften, Köln", "Köln ZB MED" ],
-  "type" : "Library",
-  "classification" : {
-    "id" : "http://purl.org/lobid/libtype#n15",
-    "type" : "Concept",
-    "label" : {
-      "de" : "Zentrale Fachbibliothek",
-      "en" : "Central Subject Library"
-    }
-  },
-  "ags" : "5315000",
-  "@@context" : "http://lobid.org/organisations/context.jsonld",
-  "url" : "http://www.zbmed.de",
-  "provides" : "http://www.livivo.de/?idb=ZBMED",
-  "name" : "Deutsche Zentralbibliothek für Medizin - Leibniz-Informationszentrum Lebenswissenschaften, Köln",
-  "containedIn" : "http://sws.geonames.org/6553049/",
-  "location" : [ {
-    "type" : "Place",
-    "address" : {
-      "postalCode" : "50931",
-      "streetAddress" : "Gleueler Str. 60",
-      "addressLocality" : "Köln",
-      "addressCountry" : "DE",
-      "type" : "PostalAddress"
-    },
-    "openingHoursSpecification" : {
-      "type" : "OpeningHoursSpecification",
-      "comment" : "Ausleihe/Allg. Information: Mo-Fr 9.00-21.00, Sa 9.00-12.00; Fachauskunft: Mo 9.00-13.00; Di-Fr 9.00-16.00; weitere Schließtage s.Homepage",
-      "description" : "Mo-Fr 9.00-21.00, Sa 9.00-17.00, So und Feiertags 12.00-17.00 (Lesesaal)"
-    },
-    "geo" : {
-      "lat" : "50.92466",
-      "lon" : "6.91666"
-    }
-  } ],
-  "isPrimaryTopicOf" : "http://dispatch.opac.d-nb.de/DB=1.2/CMD?ACT=SRCHA&IKT=8529&TRM=DE-38M",
-  "wikipedia" : "http://de.wikipedia.org/wiki/Deutsche_Zentralbibliothek_für_Medizin",
-  "id" : "http://lobid.org/organisations/DE-38M#!",
-  "isil" : "DE-38M",
-  "fundertype" : {
-    "id" : "http://purl.org/lobid/fundertype#n06",
-    "type" : "Concept",
-    "label" : {
-      "de" : "Sonstige öffentliche Trägerschaft",
-      "en" : "Other Public Sector Funder"
-    }
-  },
-  "collects" : {
-    "type" : "Collection",
-    "extent" : {
-      "id" : "http://purl.org/lobid/stocksize#n09",
-      "type" : "Concept",
-      "label" : {
-        "de" : "300.001 - 1.000.000",
-        "en" : "300,001 - 1,000,000"
-      }
-    },
-    "subject" : [ "Medizin" ]
-  },
-  "email" : "mailto:info@@zbmed.de",
-  "dbsID" : "BJ038",
-  "name_en" : "German National Library of Medicine - Leibniz Information Centre for Life Sciences, Cologne",
-  "sameAs" : [ "http://dbpedia.org/resource/Deutsche_Zentralbibliothek_für_Medizin", "http://ld.zdb-services.de/resource/organisations/DE-38M" ]
-}</code></pre></dd>
+    <dd><pre><code>@(Json.prettyPrint(Json.parse(controllers.Index.get("DE-605"))))</code></pre></dd>
     <h2 id="openrefine">OpenRefine <small><a href="#openrefine"><span class="glyphicon glyphicon-link"></span></a></small></h2>
     <p>@Html(Messages.get("api.openrefine.info", "https://github.com/OpenRefine/OpenRefine/wiki/Reconciliation-Service-API", "https://github.com/hbz/lobid-organisations/issues/55#issuecomment-129771979"))</p>
     @desc(Messages.get("api.openrefine.metadata"), routes.Reconcile.meta("jsonp"))
